Create fsl_anat
Iterate through all participants, identifying whether the file is MPRAGE or FSPGR, and then applying the fsl_anat function. Performs bias correction, brain extraction. We use -nosubcortseg and -noseg to tell the function to skip the tissue segmentation and subcortical segmentation steps. -noreg and -nonlinreg are used to skip the linear and nonlinear registration steps, so that we can do registration to different spaces later.